Moved lib to 3rdparty. Changed ContentionDecider to RevocationStrategy.
