+++ scripts/lint.sh
@@ -6,4 +6,4 @@ SERENITY_ROOT=$( dirname $(readlink -e "${DIR}/"))
 python ${DIR}/cpplint.py \
   --extensions="hpp,cpp" \
   --filter="-legal/copyright" \
-$( find "${SERENITY_ROOT}/" -name "*.cpp" -or -name "*.hpp" | grep -v -e "${SERENITY_ROOT}/build/" -e "${SERENITY_ROOT}/src/framework/" -e "${SERENITY_ROOT}/lib/" )
+$( find "${SERENITY_ROOT}/" -name "*.cpp" -or -name "*.hpp" | grep -v -e "${SERENITY_ROOT}/build/" -e "${SERENITY_ROOT}/src/framework/" -e "${SERENITY_ROOT}/3rdparty/" )
