+++ src/observers/strategies/kill_all.hpp
@@ -1,0 +1,21 @@
+#ifndef SERENITY_STRATEGIES_KILL_ALL_HPP
+#define SERENITY_STRATEGIES_KILL_ALL_HPP
+
+#include "observers/strategies/base.hpp"
+
+namespace mesos {
+namespace serenity {
+
+/**
+ * Kills all BE executors given in usage.
+ */
+class KillAllStrategy : public RevocationStrategy {
+ public:
+  RevocationStrategyFunction decide;
+};
+
+
+}  // namespace serenity
+}  // namespace mesos
+
+#endif  // SERENITY_STRATEGIES_KILL_ALL_HPP
